rf: move `get_pydantic_err_counts()`
From `cli/tools.py` to `tools.py`
